[OPR-612]: Moves NetworkObserver into ProcedureKit
Keeps iOS non-extension related elements in ProcedureKitMobile
