+++ Sources/NetworkObserver.swift
@@ -6,24 +6,12 @@
 
 import Foundation
 import Dispatch
-#if os(iOS)
-import UIKit
-#endif
 
 public protocol NetworkActivityIndicatorProtocol {
     var networkActivityIndicatorVisible: Bool { get set }
 }
 
-#if os(iOS)
-extension UIApplication: NetworkActivityIndicatorProtocol { }
-#endif
-
 public class NetworkActivityController {
-
-    #if os(iOS)
-    @available(iOSApplicationExtension, unavailable)
-    static let shared = NetworkActivityController()
-    #endif
 
     let interval: TimeInterval
     private(set) var indicator: NetworkActivityIndicatorProtocol
@@ -32,20 +20,6 @@ public class NetworkActivityController {
     private var delayedHide: DispatchWorkItem?
 
     private let queue = DispatchQueue(label: "run.kit.procedure.ProcedureKit.NetworkActivityController", qos: .userInteractive)
-
-    #if os(iOS)
-    /// (iOS-only) Initialize a NetworkActivityController that displays/hides the
-    /// network activity indicator in the status bar. (via UIApplication)
-    ///
-    /// - Parameter timerInterval: How long to wait after observed network activity stops
-    ///                            before the network activity indicator is set to false.
-    ///                            (This helps reduce flickering if you rapidly create
-    ///                            procedures with attached NetworkObservers.)
-    @available(iOSApplicationExtension, unavailable, message: "Not supported in Application Extensions because UIApplication.shared is unavailable. Use init(indicator:) or init(timerInterval:indicator:) instead.")
-    public convenience init(timerInterval: TimeInterval = 1.0) {
-        self.init(timerInterval: timerInterval, indicator: UIApplication.shared)
-    }
-    #endif
 
     /// Initialize a NetworkActivityController
     ///
@@ -114,15 +88,6 @@ public class NetworkObserver: ProcedureObserver {
     public init(controller: NetworkActivityController) {
         networkActivityController = controller
     }
-
-    #if os(iOS)
-    /// (iOS-only) Initialize a NetworkObserver that displays/hides
-    /// the network activity indicator in the status bar. (via UIApplication)
-    @available(iOSApplicationExtension, unavailable, message: "Not supported in Application Extensions because UIApplication.shared is unavailable. Use init(controller:) instead.")
-    public convenience init() {
-        self.init(controller: NetworkActivityController.shared)
-    }
-    #endif
 
     public func will(execute procedure: Procedure) {
         networkActivityController.start()
